requirements.txt
python-telegram-bot==21.4
yt-dlp==2024.8.6
python-dotenv==1.0.1
